# WAL: store pointers in record pool