**What this PR does / why we need it**:

Storing a slice in `sync.Pool` wastes 24 bytes every time it is returned to the heap; see https://staticcheck.io/docs/checks#SA6002

Storing a pointer to a slice allows the Pool to work without waste.

This came up in discussion about the benchmark in #8489 - with this change the memory usage comes down:
```
name                                                      old time/op    new time/op    delta
Writer_WriteEntries/1000_lines,_1_different_label_sets-4    7.37ms ±13%    7.45ms ± 9%     ~     (p=0.841 n=5+5)

name                                                      old alloc/op   new alloc/op   delta
Writer_WriteEntries/1000_lines,_1_different_label_sets-4     727kB ± 0%     679kB ± 0%   -6.63%  (p=0.008 n=5+5)

name                                                      old allocs/op  new allocs/op  delta
Writer_WriteEntries/1000_lines,_1_different_label_sets-4     12.0k ± 0%     10.0k ± 0%  -16.68%  (p=0.000 n=5+4)
```

**Checklist**
- [x] Reviewed the [`CONTRIBUTING.md`](https://github.com/grafana/loki/blob/main/CONTRIBUTING.md) guide (**required**)
- NA Documentation added
- [x] Tests updated
- NA `CHANGELOG.md` updated
- NA Changes that require user attention or interaction to upgrade are documented in `docs/sources/upgrading/_index.md`